# backend/docs/swagger.yaml
basePath: /
definitions:
  filesystem.PathValidationResult:
    properties:
      error:
        type: string
      exists:
        type: boolean
      isDir:
        type: boolean
      readable:
        type: boolean
      validated:
        type: boolean
      writable:
        type: boolean
    type: object
...
  models.Workspace:
    properties:
      createdAt:
        type: string
      id:
        type: string
      lastOpenedAt:
        type: string
      name:
        type: string
      path:
        type: string
    type: object
  models.WorkspaceCreate:
    properties:
      name:
        type: string
      ownerDisplayName:
        type: string
      path:
        type: string
    required:
    - name
    - path
    type: object
  models.WorkspaceUpdate:
    properties:
      name:
        type: string
      path:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Multi-Agent Collaboration Platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Orchestra API
  version: "1.0"
paths:
...
  /api/browse:
    get:
      description: Browse files from home directory or specified path
      parameters:
      - description: Path to browse
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Browse root files
      tags:
      - workspaces
...
  /api/workspaces:
    get:
      description: Get all workspaces
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Workspace'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List workspaces
      tags:
      - workspaces
    post:
      consumes:
      - application/json
      description: Create a new workspace
      parameters:
      - description: Workspace data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.WorkspaceCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Workspace'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create workspace
      tags:
      - workspaces
  /api/workspaces/{id}:
    delete:
      description: Delete a workspace by ID
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete workspace
      tags:
      - workspaces
    get:
      description: Get a workspace by ID
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Workspace'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get workspace
      tags:
      - workspaces
    put:
      consumes:
      - application/json
      description: Update workspace name or other properties
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Workspace update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.WorkspaceUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Workspace'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update workspace
      tags:
      - workspaces
...
  /api/workspaces/{id}/browse:
    get:
      description: Browse files in a workspace directory
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Sub path to browse
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Browse workspace files
      tags:
      - workspaces
...
  /api/workspaces/{id}/search:
    get:
      description: Full-text search across all conversations in a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 50
        description: Maximum results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Search workspace messages
      tags:
      - workspaces
...
  /api/workspaces/validate-path:
    post:
      consumes:
      - application/json
      description: Validate if a path is suitable for a workspace (exists, readable,
        writable)
      parameters:
      - description: Path to validate
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/filesystem.PathValidationResult'
      security:
      - BearerAuth: []
      summary: Validate workspace path
      tags:
      - workspaces
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"